Move executables to cmd/<program>.
- Adjust scripts to work properly;
- Remove unused scripts;

Closes #10
