+++ Dockerfile
@@ -2,5 +2,5 @@ FROM scratch
 
 MAINTAINER Guilherme Santos <guilherme.santos@neoway.com.br>
 
-ADD ./machine-up /opt/cloud-machine/bin/
-ADD ./cluster-up /opt/cloud-machine/bin/
+ADD ./cmd/machine-up/machine-up /opt/cloud-machine/bin/
+ADD ./cmd/cluster-up/cluster-up /opt/cloud-machine/bin/
