+++ Makefile
@@ -1,21 +1,27 @@
-version ?= latest
+version ?= 1.0
 WORKDIR="github.com/NeowayLabs/cloud-machine"
 IMAGENAME=neowaylabs/cloud-machine
 IMAGE=$(IMAGENAME):$(version)
 
 all: machine-up cluster-up
-	@echo "Create: machine-up & cluster-up"
+	@echo "Created: machine-up & cluster-up"
 
 goget:
 	go get -d -v ./...
 
-machine-up: goget
-	go build -v machine-up.go auth.go
+machine-up:
+	cd cmd/machine-up && make build
 
-cluster-up: goget
-	go build -v cluster-up.go auth.go
+cluster-up:
+	cd cmd/cluster-up && make build
 
-install: deploy
+build: machine-up cluster-up
+
+static:
+	cd cmd/machine-up && make static
+	cd cmd/cluster-up && make static
+	ldd cmd/machine-up/machine-up | grep "not a dynamic executable"
+	ldd cmd/cluster-up/cluster-up | grep "not a dynamic executable"
 
 publish: image
 	docker push $(IMAGE)
@@ -23,15 +29,11 @@ publish: image
 image: build
 	docker build -t $(IMAGE) .
 
-build: build-env
-	docker run --rm -v `pwd`:/go/src/$(WORKDIR) --privileged -i -t $(IMAGENAME) bash hack/make.sh
+build-static: build-env
+	docker run --rm -v `pwd`:/go/src/$(WORKDIR) --privileged -i -t $(IMAGENAME) make static
 
 build-env:
 	docker build -t $(IMAGENAME) -f ./hack/Dockerfile .
 
-check: build
-	docker run --rm -v `pwd`:/go/src/$(WORKDIR) --privileged -i -t $(IMAGENAME) bash hack/check.sh
-
-shell: build
+shell: build-env
 	docker run --rm -v `pwd`:/go/src/$(WORKDIR) --privileged -i -t $(IMAGENAME) bash
-
